Create app directory move app component, Fix app import in indexjs
