+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './Components/App';
+import App from './Components/App/App';
 import * as serviceWorkerRegistration from './Services/Workers/serviceWorkerRegistration';
 import reportWebVitals from './Services/WebVitals/reportWebVitals';
